roundf() ipv round()

+++ module02/ex01/src/Fixed.cpp
@@ -35,7 +35,7 @@ Fixed::Fixed( const float float_value) : _fixed_point(0) {
 
 	whole = (int)float_value;
 	fractional = float_value - (float)whole;
-	this->setRawBits((whole << _fractional_bits) + (int)round((fractional * (1 << _fractional_bits))));
+	this->setRawBits((whole << _fractional_bits) + (int)roundf((fractional * (1 << _fractional_bits))));
 	std::cout << "Float constructor called" << std::endl;
 
 }
